// src/shared/api/generated/models/examinationManageContractDtoSubjectGetSubjectsByGroupInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */

export interface ExaminationManageContractDtoSubjectGetSubjectsByGroupInput {
  subjectGroupType?: number;
  subjectGroupId?: string;
  examTypeId?: string;
  index?: number;
  size?: number;
}