// src/shared/api/generated/models/examinationManageContractDtoSubjectGetSubjectsInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */

export interface ExaminationManageContractDtoSubjectGetSubjectsInput {
  /** 是否今天 */
  isToday?: boolean;
  /** 1判断2单选3多选 */
  type?: string;
  /** 已做/未做 */
  haveDone?: boolean;
  /** 传1查询易错题 */
  easyLevel?: number;
  /** 科目下的子科目Id */
  examTypeId?: string;
  /** 科目ID */
  examTypeParentId?: string;
  /** 是否为错题 */
  isMistake?: boolean;
  /** 是否为收藏题 */
  isCollection?: boolean;
  subjectGroupId?: string;
  index?: number;
  size?: number;
}